requirements.txt
# Core Dependencies
pathway>=0.8.0
python-dotenv>=1.0.0

# BDH Architecture
torch>=2.0.0
transformers>=4.35.0
einops>=0.7.0

# NLP & Text Processing
nltk>=3.8.1
spacy>=3.7.0
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0

# Utilities
pyyaml>=6.0
loguru>=0.7.0
